// src/services/api/generated/types.gen.ts
export type SearchRequest = {
    query: string;
    projectId: Uuid;
};

export type SearchResponse = {
    results?: Array<SearchResult>;
};

export type SearchResult = {
    id?: Uuid;
    chunkText?: string;
    chunkIndex?: number;
    source?: string;
    distance?: number;
};

export type TextRequest = {
    text: string;
    projectId: Uuid;
};

export type Uuid = string;